Rework remote tests to avoid issues.
