+++ trippi-sail-blazegraph-remote/src/test/java/ca/discoverygarden/trippi/sesame/blazegraph/RemoteBlazegraphViaSparqlIntegrationTest.java
@@ -3,6 +3,6 @@ package ca.discoverygarden.trippi.sesame.blazegraph;
 import org.springframework.test.context.ContextConfiguration;
 
 @ContextConfiguration
-public class RemoteBlazegraphViaSparqlIntegrationTest extends RemoteBlazegraphIntegrationTest {
-	// No-op; everything is in RemoteBlazegraphIntegrationTest.
+public class RemoteBlazegraphViaSparqlIntegrationTest extends AbstractRemoteBlazegraphIntegrationTest {
+	// No-op; everything is in AbstractRemoteBlazegraphIntegrationTest.
 }
